% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_mask.R
\name{binary_mask}
\alias{binary_mask}
\title{binary_mask}
\usage{
binary_mask(binary_mask, clipped, output_folder)
}
\arguments{
\item{binary_mask}{The raster data of the binary mask.}

\item{clipped}{list.files("path/to/raster/files", pattern = "\\.tif$", full.names = TRUE)}

\item{output_folder}{<- "path/to/output/directory"}
}
\value{
Nothing. The results are saved in the specified directory.
}
\description{
#Apply a binary mask to raster data
#All data in Global Snow pack is stored as 8-bit unsigned integers.
#he binary mask is created by setting all values below 64 to 0 and all values above or equal to 64 to 1.
}
\details{
#This function takes a binary mask and a list of raster data to which the mask should be applied.
#It applies the binary mask to all the rasters and saves the results.
}
\examples{
binary_mask(binary_mask, clipped, "path/to/output/directory")

}
