% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_scene_snowstats.R
\name{calc_scene_snowstats}
\alias{calc_scene_snowstats}
\title{calc_scene_snowstats}
\usage{
calc_scene_snowstats(store_directory, dem_subdir, height_zone_size,
  threshold_cloud, last_snow = FALSE)
}
\arguments{
\item{store_directory}{Directory where all satellite scenes are stored in subfolders}

\item{dem_subdir}{subdirectory, where DEM files are stored}

\item{height_zone_size}{defines the size (height difference) of the altitudinal zones for which the snowcover percentages are calculated}

\item{thresholds_cloud}{thresholds for cloud cover to decide, whether the scene will be used in time series or not}
}
\description{
This function named 'calc_scene_snowstats' loads the snow/cloud/freesurface classification tif file and calculates cover percentages.
\preformatted{
This is done for the whole testsite, to decide, whether the scene will be used
for the timeseries or not. Used Thresholds for decision:
     - cloud cover less than input threshold
     - cloud cover less than freesurface
 => Decision is stored as TRUE or FALSE in a text-file in the satellite scene subfolder.}

Additionally snow/cloud/freesurface cover percentages are also calculated for each height zone derrived from DEM file and stored as table in satellite scene subfolder.
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
