% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_snowextent.R
\name{calculate_snowextent}
\alias{calculate_snowextent}
\title{calculate_snowextent}
\usage{
calculate_snowextent(store_directory, s02, ls8, thresholds_snow_cloud)
}
\arguments{
\item{store_directory}{Directory where all satellite scenes are stored in subfolders}

\item{s02, ls8}{defines, whether data from Sentinel or Landsat subfolder is used}

\item{thresholds_snow_cloud}{set thresholds for snow extent & cloud cover for the resulting classification.
\preformatted{
The parameter contains a list of 4 thresholds
first:   NDSI minimum value for snowcover
second:  minimum green reflectance for snowcover
third:   minimum green reflectance for cloudcover
fourth:  minimum SWIR reflectance for cloudcover}}
}
\description{
This function named 'calculate_snowextent' loads testsite SWIR & green tif files and calculates NDSI values.
Additionally a simple snow/no-snow classification and a snow/cloud/snowfree-surface classification is produced.
\preformatted{
Therefore thresholds are used:
Snow: minimum NDSI, minimum reflectance in green band
Cloud: minimum reflectance in green band and SWIR band + existence of snow}
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
