% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_images_snow.R
\name{create_images_snow}
\alias{create_images_snow}
\title{create_images_snow}
\usage{
create_images_snow(store_directory, images, RGBimage_bands = FALSE,
  new_scenes = FALSE)
}
\arguments{
\item{store_directory}{Directory where all satellite scenes are stored in subfolders}

\item{images}{vector with strings that defines which images are calculated
\preformatted{ Possible input strings:
    - 'snow':   plots snow/no-snow classification extent map
    - 'snow_cloud': plots snow/cloud/freesurface classification map
    - 'new_snow': plots recalculated snow extend from new_snow.tif file
             (produced only for timeseries used data; look at function 'calc_scene_snowstats')
    - 'RGB': plot RGB false or true color from the three bands in variable 'RGBimage_bands'
    - 'NDSI': plots a greyscale image of NDSI}}

\item{RGBimage_bands}{vector containing 3 strings with bands, which should be useed for RGB imaging}

\item{new_scenes}{set true if you want only RGB-images to be plotted, where 'new_snow' extent was calculated. A new subfolder will be created.}
}
\description{
This function named 'create_images_snow' loads three tif files for RGB image or NDSI tif file or
classification tif files to plot them.
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
