% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow_animation.R
\name{create_snow_animation}
\alias{create_snow_animation}
\title{snow_animation}
\usage{
create_snow_animation(raster_files_snow, date, country_name)
}
\arguments{
\item{raster_files_snow}{A character vector containing the file paths of raster layers representing snow cover for each day}

\item{date}{A character vector of dates corresponding to each raster image}

\item{Country_name}{A character string specifying the name of the country}
}
\value{
Nothing. The GIF animation is saved as "snow_animation.gif"
}
\description{
#Create a GIF animation of snow cover changes over time
#This function creates a GIF animation illustrating the changes in snow cover over time using a series of raster files
}
\details{
This function takes a list of raster files representing snow cover for each date and creates a GIF animation
to visualize the changes in snow cover over time. The animation displays one frame for each date, with each frame
showing the snow cover image for that date
}
\examples{
create_snow_animation(raster_files_snow, date, "country_name")

}
