% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_rasters.R
\name{crop_rasters}
\alias{crop_rasters}
\title{crop_rasters}
\usage{
crop_rasters(raster_files, country_boundary, output_dir)
}
\arguments{
\item{raster_files}{A character vector containing paths to the raster files to be cropped}

\item{country_boundary}{The boundary of the country as a SpatialPolygonsDataFrame or sf object}

\item{output_dir}{The directory where the cropped raster files will be saved}
}
\value{
None
}
\description{
#This function crops raster files based on the boundary of a country
}
\examples{
raster_files <- list.files("path/to/raster/files", pattern = "\\\\.tif$", full.names = TRUE)
output_dir <- "path/to/output/directory"
crop_rasters(raster_files, country_boundary, output_dir)

}
