% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tarlist.R
\name{data_tarlist}
\alias{data_tarlist}
\title{data_tarlist}
\usage{
data_tarlist(input_dir, input_file)
}
\arguments{
\item{input_dir}{Folder directory to the tar.gz file}

\item{inputfile}{name of tar.gz-file}
}
\description{
is returning a list of all the files a tar.gz-file contains
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
