% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_untar.R
\name{data_untar}
\alias{data_untar}
\title{data_untar}
\usage{
data_untar(input_dir, input_file, output_dir, subfile)
}
\arguments{
\item{input_dir}{Folder directory to tar.gz file}

\item{output_dir}{directory, where untarred files should be stored}

\item{subfile}{get only the files out of the tar.gz-file, that are needed}

\item{inputfile}{name of tar.gz-file}
}
\description{
is untarring files
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
