% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_unzip.R
\name{data_unzip}
\alias{data_unzip}
\title{data_unzip}
\usage{
data_unzip(input_dir, input_file, output_dir, subfile)
}
\arguments{
\item{input_dir}{Folder directory to zip file}

\item{output_dir}{directory, where unzipped files should be stored}

\item{subfile}{get only the files out of the zip-file, that are needed}

\item{inputfile}{name of zip-file}
}
\description{
is unzipping files
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
