% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ziplist.R
\name{data_ziplist}
\alias{data_ziplist}
\title{data_ziplist}
\usage{
data_ziplist(input_dir, input_file)
}
\arguments{
\item{input_dir}{Folder directory to zip file}

\item{inputfile}{name of zip-file}
}
\description{
is returning a list of all the files a zip-file contains
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
