% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_snow_data.R
\name{download_snow_data}
\alias{download_snow_data}
\title{Download Snow Data}
\usage{
download_snow_data(start_date, end_date, save_dir)
}
\arguments{
\item{start_date}{The start date in the format as.Date("YYYY-MM-DD")}

\item{end_date}{The end date in the format as.Date("YYYY-MM-DD")}

\item{save_dir}{The directory path where the downloaded files will be saved.}
}
\value{
Nothing. The images are saved to the specified directory.
}
\description{
#This function downloads daily snow cover images from the specified start date to the specified end date.
}
\details{
#This function generates URLs for downloading snow cover data for each date in the specified range.
}
\examples{
start_date <- as.Date("2023-01-01")
end_date <- as.Date("2023-01-31")
download_snow_data(start_date, end_date, "path/to/save/directory")

}
