% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Landsat8_data_from_tar.R
\name{get_Landsat8_data_from_tar}
\alias{get_Landsat8_data_from_tar}
\title{get_Landsat8_data_from_tar}
\usage{
get_Landsat8_data_from_tar(data_dir, dir_aim, used_band, add = TRUE)
}
\arguments{
\item{data_dir}{Folder directory to tar files}

\item{dir_aim}{Folder, where extracted files / folder structure should be stored}

\item{used_band}{string vector with bands, that should be extractet (e.g "band1", "band2",..., "band7")}

\item{add}{if variable is set TRUE, new layer directories will be added to already existing directory file
(mandatory that directory file already exists)}
}
\description{
This function extracts all needed Landsat8 files and stores them in subfolder of the aimfolder structure.

A txt file is provided that contains the paths to the extracted files
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
