% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Sentinel2_data_from_zip.R
\name{get_Sentinel2_data_from_zip}
\alias{get_Sentinel2_data_from_zip}
\title{get_Sentinel2_data_from_zip}
\usage{
get_Sentinel2_data_from_zip(data_dir, dir_aim, used_band, used_res,
  add = TRUE)
}
\arguments{
\item{data_dir}{Folder directory to zip files}

\item{dir_aim}{Folder, where extracted files / folder structure should be stored}

\item{used_band}{string vector with bands, that should be extractet (e.g "B02", "B11", "AOT", "SCL")}

\item{used_res}{string vector with corresponding resolution ("10", "10",...), must be of same size as used_band,

Check for available Resolution beforehand!!!}

\item{add}{if variable is set TRUE, new layer directories will be added to already existing directory file
(mandatory that directory file already exists)}
}
\description{
This function named 'get_Sentinel2_data_from_zip' extracts all needed S2 files and stores them in their folder structure.

A txt file is provided that contains the paths to the extracted files
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
