% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_testsite_Data.R
\name{get_testsite_Data}
\alias{get_testsite_Data}
\title{get_testsite_Data}
\usage{
get_testsite_Data(store_directory, s02, ls8, ref_CRS, ext, res, needed_bands,
  name_bands)
}
\arguments{
\item{store_directory}{Directory where all satellite scenes are stored in subfolders}

\item{s02, ls8}{defines, whether data from Sentinel or Landsat subfolder is used}

\item{ref_CRS}{used coordinate reference system; in case data is stored in different coordinate systems,
it will be reprojected to this coordinate system}

\item{ext}{extent of testsite the scenes will be cropped to}

\item{res}{if data needs to be reprojected, this will be the resolution of the resulting raster file}

\item{needed_bands}{define bands, which should be cropped (these are the input file names - satellite specific)
(e.g "band1", "band5", "B04", "B08")}

\item{name_bands}{define name of output tif-files (e.g "green", "NIR", "SWIR")}
}
\description{
This function named 'get_testsite_Data' loads entire satellite scene tif files of selected bands and
crops them to a predefined extent, the result is stored in a new subfolder 'testsite' in each acquisition folder
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
