% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_country_boundary.R
\name{load_country_boundary}
\alias{load_country_boundary}
\title{load_country_boundary}
\usage{
load_country_boundary(map, country_name)
}
\arguments{
\item{map}{A spatial object containing country boundaries, such as a 'sf' object}

\item{country_name}{The name of the country whose boundary is to be loaded}
}
\value{
A geometry object representing the boundary of the specified country
}
\description{
#Load the boundary of a specific country from a map object
#This function extracts the boundary geometry of a specific country from a map
#object and returns it
}
\examples{
library(sf)
map <- ne_download(scale = 10, returnclass = "sf")
country_boundary <- load_country_boundary(map, "Denmark")
plot(country_boundary)

}
