% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_snow_area.R
\name{mean_snow_cover}
\alias{mean_snow_cover}
\title{mean_snow_cover}
\usage{
mean_snow_cover(raster_files_snow)
}
\arguments{
\item{raster_files_snow}{A list of raster files containing snow cover data}
}
\value{
The mean snow cover value
}
\description{
#Calculate the mean snow cover image from a stack of raster files
}
\details{
#This function takes a list of raster files containing snow cover data and calculates
#the mean snow cover image across all files in the stack
}
\examples{
mean_image <- mean_snow_cover(raster_files_snow)

}
