% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snow_cover.R
\name{plot_snow_cover}
\alias{plot_snow_cover}
\title{plot_snow_cover}
\usage{
plot_snow_cover(snow_cover_area, dates, country_name)
}
\arguments{
\item{snow_cover_area}{A numeric vector containing the snow cover area data in square kilometers for each day.}

\item{dates}{A vector of class Date containing the dates corresponding to each day.}

\item{country_name}{A character string specifying the name of the country.}
}
\value{
A ggplot object displaying the snow cover changes over time.
}
\description{
#Plot Snow Cover Changes Over Time
#This function plots the changes in snow cover over time for a specified country using hystogram and trend line
}
\examples{
snow_cover_area <- unlist(snow_cover_area_list)
dates <- seq(as.Date("2023-01-01"), by = "day", length.out = nlayers(snow_cover_series))
country_name <- "Denmark"
plot_snow_cover(snow_cover_area, dates, country_name)

}
