% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_snowextent_scene.R
\name{recalc_snowextent_scene}
\alias{recalc_snowextent_scene}
\title{recalc_snowextent_scene}
\usage{
recalc_snowextent_scene(store_directory, dem_subdir, height_zone_size,
  threshold1_snow, threshold2_cloud)
}
\arguments{
\item{store_directory}{Directory where all satellite scenes are stored in subfolders}

\item{dem_subdir}{subdirectory, where DEM files are stored}

\item{height_zone_size}{defines the size (height difference) of the altitudinal zones for which the snowcover percentages are calculated;
must be the same size as used in function 'calc_scene_snowstats'}

\item{threshold1_snow}{set thresholds for heigth zone snow probability, where pixel will be excluded from the class snow}

\item{threshold2_cloud}{set thresholds for heigth zone snow probability, above which cloud pixels will be added to the class snow}
}
\description{
This function named 'recalc_snowextent_scene' recalculates snow extent to exclude low probable snow pixels and
include high probable snow under cloudcover. For that the previously calculated dem height zone probabilities are used
(function 'calc_scene_snowstats')
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
