% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_3d_snow.R
\name{results_3d_snow}
\alias{results_3d_snow}
\title{results_3d_snow}
\usage{
results_3d_snow(store_directory, subfolder_img = FALSE, dem_subdir,
  texture_file, locations = FALSE, views, color_snow, color_locs)
}
\arguments{
\item{store_directory}{Directory where all satellite scenes are stored in subfolders}

\item{dem_subdir}{subdirectory, where DEM files are stored}

\item{texture_file}{defines png-file, which should be used as background for 3d plot (must be exactly fitting to testsite scene & mirrored horizontally; e.g. with gimp)}

\item{locations}{if a csv-file with locations within testsite exits, add it here so that their locations and names will be added to the image
\preformatted{ Example:
name;E;N;height;UTM
Innsbruck;681613;5236564;600;32
Sterzing;685358;5196180;950;32
Brenner;690605;5208832;1380;32}}

\item{views}{decide which views should be plotted; "north", "south" and "top" are possible}
}
\description{
his function named 'results_3d_snow' recalculated snow extend tif files and plots them as red overlay in 3D on a RGB-image useing the DEM raster.
Several point of views can be produced, but is needed to be adjusted for different testsites.
If a locations csv-file is added remarkable points in the scene can be plotted and named in the image for better recognition of scene.
}
\author{
Sebastian Buchelt (M.Sc. EAGLE - University of Wuerzburg)
}
